Refactoring to move wrapping code/tests to src/plugins directory, rename redis wrapper to ioredis (anticipating other redis wrappers), shortening redis env var, specifying ioredis in env var.
